requirements.txt
streamlit
sqlalchemy
psycopg2-binary
google-generativeai
pandas
python-dotenv
requests
markdown
reportlab